requirements.txt
# Root requirements file for Railway deployment
# Installs all dependencies needed for CleanoutPro backend

# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
psycopg[binary]>=3.2.0
alembic==1.13.1

# Redis
redis==5.0.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client for external APIs
requests==2.31.0
httpx==0.26.0

# Image processing
Pillow==10.2.0
aiofiles==23.2.1

# PDF generation for invoices
reportlab==4.0.9

# WebSocket support
websockets==12.0
python-socketio==5.10.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dateutil==2.8.2

# Logging
structlog==24.1.0
